backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pymupdf>=1.22.5
langchain>=0.2.0
faiss-cpu>=1.7.4
tiktoken>=0.4.0
python-dotenv>=1.0.0
supabase>=1.0.0
httpx>=0.24.0
pydantic>=1.10.7
sqlalchemy>=2.0.15
aiofiles>=23.1.0
langchain-google-genai>=0.0.6